Refactor frontend into modules with service layer and tests
