+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+  "semi": true,
+  "singleQuote": false,
+  "printWidth": 100
+}
