+++ eslint.config.js
@@ -1,0 +1,22 @@
+import globals from "globals";
+
+export default [
+  {
+    files: ["**/*.js"],
+    languageOptions: {
+      sourceType: "module",
+      ecmaVersion: "latest",
+      globals: {
+        ...globals.browser,
+        ...globals.node
+      }
+    },
+    rules: {
+      "no-unused-vars": ["error", { "argsIgnorePattern": "^_" }],
+      "no-undef": "error"
+    }
+  },
+  {
+    ignores: ["node_modules/**"]
+  }
+];
