+++ firebase.js
@@ -1,0 +1,11 @@
+export {
+  getPaths,
+  subscribe,
+  tasksApi,
+  notesApi,
+  financeApi,
+  dailyTasksApi,
+  habitsApi,
+  focusApi,
+  activityApi
+} from "./firebaseService.js";
