+++ index.html
@@ -85,6 +85,6 @@
     </section>
   </div>
 
-  <script type="module" src="scripts.js"></script>
+  <script type="module" src="ui.js"></script>
 </body>
 </html>
