+++ package.json
@@ -1,0 +1,18 @@
+{
+  "name": "i-can-do",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "test": "node --test",
+    "lint": "eslint .",
+    "format": "prettier --check .",
+    "format:write": "prettier --write .",
+    "dev": "vite"
+  },
+  "devDependencies": {
+    "eslint": "^9.20.1",
+    "globals": "^15.14.0",
+    "prettier": "^3.5.1",
+    "vite": "^6.2.1"
+  }
+}
